+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Tantivy 0.11.0
 - Avoid rebuilding Regex automaton whenever a regex query is reused. #639 (@brainlock)
 - Add footer with some metadata to index files. #605 (@fdb-hiroshima)
 - TopDocs collector: ensure stable sorting on equal score. #671 (@brainlock)
+- Fix crash when committing multiple times with deleted documents. #681 (@brainlock)
  
 ## How to update?